requirements.txt
torch
torchvision
pytorch-lightning
tqdm
pandas
wandb
hydra-core
boto3
